+++ src/budge/account.py
@@ -1,10 +1,12 @@
 from dataclasses import dataclass, field
-from datetime import date
+from datetime import date, timedelta
 from heapq import merge
 from itertools import groupby
+from typing import Generator
 
 from stockholm import Money
 
+from .date import daterange
 from .transaction import RepeatingTransaction, Transaction
 
 
@@ -47,27 +49,61 @@ class Account:
             for transaction in self.transactions_range(end_date=as_of)
         )
 
-    def balance_iter(
+    def daily_balance(
         self, start_date: date | None = None, end_date: date | None = None
-    ):
+    ) -> Generator[tuple[date, Money]]:
         """
-        Iterate over the account's balance over the given range, yielding a
-        tuple of each date in the range and the account balance on that date.
+        Iterate over the daily balance of the account, yielding tuples of date
+        and balance.
 
-        If `start_date` is not provided, the first yield will be the initial
-        balance of the account.
-
-        If `end_date` is not provided, the iteration will continue until all
-        transactions in the account have been iterated over.
-
-        :param start_date: The start date of the range
-        :param end_date: The end date of the range
-        :yield: A tuple of (date, balance) for each day in the range
+        The balance on the given start date is yielded first, and then the
+        balance for each subsequent date is yielded. If the start date is not
+        given, the date of the first transaction is used. If the end date is not
+        given, today's date is used.
         """
-        bal = self.balance(start_date) if start_date else Money(0)
+        start_date = start_date or next(self.transactions_range()).date
+        end_date = end_date or date.today()
 
-        for date_, transactions in groupby(
-            self.transactions_range(start_date, end_date), lambda t: t.date
+        balance = self.balance(start_date)
+        yield start_date, balance
+
+        for _date, delta in self._daily_balance_delta(
+            start_date + timedelta(days=1), end_date
         ):
-            bal += Money.sum(transaction.amount for transaction in transactions)
-            yield date_, bal
+            balance += delta
+            yield _date, balance
+
+    def _deltas_by_date(self, start_date: date, end_date: date):
+        """
+        Iterate over the deltas in the account balance for each date in the
+        given range, including the given start and end dates.
+
+        Yields tuples, where the first element is the date and the second
+        element is the total amount of all transactions on that date.
+        """
+        return (
+            (_date, Money.sum(transaction.amount for transaction in transactions))
+            for _date, transactions in groupby(
+                self.transactions_range(start_date, end_date), key=lambda t: t.date
+            )
+        )
+
+    def _daily_balance_delta(self, start_date: date, end_date: date):
+        """
+        Calculate the daily change in account balance over the specified date range.
+
+        This function yields tuples, where the first element is the date and the
+        second element is the net change in balance for that date. It combines the
+        deltas from actual transactions and placeholder deltas for dates without
+        transactions to ensure a continuous range.
+        """
+        return (
+            (_date, Money.sum(delta[1] for delta in deltas))
+            for _date, deltas in groupby(
+                merge(
+                    self._deltas_by_date(start_date, end_date),
+                    ((_date, Money(0)) for _date in daterange(start_date, end_date)),
+                ),
+                key=lambda t: t[0],
+            )
+        )